delete unnecessary files add permission control to some routes make all repeated responses functions

+++ shtl_ink/shtl_ink_api/app.py
@@ -74,12 +74,39 @@ def json_response_in_use(short_code):
         {"message": f"{short_code} already in use"},
         status_code=status.HTTP_409_CONFLICT)
 
+def json_response_not_owned(short_code):
+    return JSONResponse(
+        {"message": f"{short_code} not owned by you"},
+        status_code=status.HTTP_403_FORBIDDEN)
 
 def json_response_created(url_record):
     return JSONResponse(
         url_record.to_dict(),
         status_code=status.HTTP_201_CREATED)
 
+def json_response_already_reported(url_record):
+    return JSONResponse(
+            url_record.to_dict(),
+            status_code=status.HTTP_208_ALREADY_REPORTED)
+
+def json_response_deleted(short_code, url):
+    return JSONResponse(
+            {"message": f"deleted record {short_code} -> {url}"},
+            status_code=status.HTTP_200_OK)
+
+def json_response_record(url_record):
+    return JSONResponse(
+            url_record.to_dict(),
+            status_code=status.HTTP_200_OK)
+
+def json_response_missing(something):
+    return JSONResponse({"message": f"you must supply {something}"},
+                            status_code=status.HTTP_406_NOT_ACCEPTABLE)
+
+def json_response_failure():
+    return JSONResponse(
+            {"message": "something went wrong..."},
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
 # all reserved endings that have no form data need to be before
 # short code redirect reciever
 
@@ -97,7 +124,6 @@ async def get_all_records(
     session: SessionContainer = Depends(verify_session(session_required=False))):
 
     user_id = get_user_id(session)
-    print(f"USER: {user_id} requesting all short codes...")
     url_records = db.execute(select(ShortURLModel).where(ShortURLModel.owner_id == user_id)).scalars().all()
     url_records = [url_record.to_dict() for url_record in url_records]
     return JSONResponse(url_records)
@@ -125,13 +151,11 @@ async def create_short_code(
         create_request: CreateRequest,
         db: Session = Depends(get_db),
         session: SessionContainer = Depends(verify_session(session_required=False))):
-    
+
     user_id = get_user_id(session)
-    print(f"USER: {user_id} creating short code...")
 
     if create_request.url == '':
-        return JSONResponse({"message": "you must supply a url"},
-                            status_code=status.HTTP_406_NOT_ACCEPTABLE)
+        return json_response_missing("a url")
 
     url_record = db.execute(
         select(ShortURLModel).where(
@@ -141,37 +165,38 @@ async def create_short_code(
         short_code = codec.encode(create_request.url, user_id, db)
 
     else:
-        return JSONResponse(url_record.to_dict(),
-                            status_code=status.HTTP_208_ALREADY_REPORTED)
+        return json_response_already_reported(url_record)
 
     if isinstance(short_code, str):
         url_record = db.get(ShortURLModel, (short_code))
         return json_response_created(url_record)
 
     else:
-        return JSONResponse(
-            {"message": "something went wrong..."},
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        return json_response_failure()
 
 
 @app.post("/create_custom_short_code")
 async def create_custom_short_code(
         create_custom_request: CreateCustomRequest,
-        db: Session = Depends(get_db)):
+        db: Session = Depends(get_db),
+        session: SessionContainer = Depends(verify_session(session_required=False))):
+
+    user_id = get_user_id(session)
 
     if create_custom_request.short_code == '' or create_custom_request.url == '':
-        return JSONResponse({"message": "you must supply a url and short code"},
-                            status_code=status.HTTP_406_NOT_ACCEPTABLE)
+        return  json_response_missing("a url and short code")
 
     url_record = db.get(ShortURLModel, (create_custom_request.short_code))
 
-    if url_record is not None and url_record.url == create_custom_request.url:
-        return JSONResponse(
-            url_record.to_dict(),
-            status_code=status.HTTP_200_OK)
-
+    if url_record is not None and url_record.url == create_custom_request.url and url_record.owner_id == user_id:
+        return json_response_already_reported(url_record)
+    
+    if url_record is not None:
+        return json_response_in_use(url_record.short_code)
+    
     try:
         url_record = ShortURLModel(
+            owner_id=user_id,
             url=create_custom_request.url,
             short_code=create_custom_request.short_code)
         db.add(url_record)
@@ -189,8 +214,7 @@ async def Get_short_code_url(
         db: Session = Depends(get_db)):
 
     if url_request.short_code == '':
-        return JSONResponse({"message": "you must supply a short code"},
-                            status_code=status.HTTP_406_NOT_ACCEPTABLE)
+        return json_response_missing("a short code")
 
     url_record = db.get(ShortURLModel, (url_request.short_code))
 
@@ -198,9 +222,7 @@ async def Get_short_code_url(
         return json_response_not_found(url_request.short_code)
 
     else:
-        return JSONResponse(
-            url_record.to_dict(),
-            status_code=status.HTTP_200_OK)
+        return json_response_record(url_record)
 
 
 @app.get("/short_code/{short_code}")
@@ -213,81 +235,86 @@ async def get_short_code_url(
         return json_response_not_found(short_code)
 
     else:
-        return JSONResponse(
-            url_record.to_dict(),
-            status_code=status.HTTP_200_OK)
+        return json_response_record(url_record)
 
 
 @app.delete("/delete_short_code")
 async def Delete_url_short_code(
         url_request: UrlRequest,
-        db: Session = Depends(get_db)):
+        db: Session = Depends(get_db),
+        session: SessionContainer = Depends(verify_session(session_required=False))):
+
+    user_id = get_user_id(session)
 
     if url_request.short_code == '':
-        return JSONResponse({"message": "you must supply a short code"},
-                            status_code=status.HTTP_406_NOT_ACCEPTABLE)
+        return json_response_missing("a short code")
 
     url_record = db.get(ShortURLModel, (url_request.short_code))
 
     if url_record is None:
         return json_response_not_found(url_request.short_code)
+    
+    if url_record.owner_id != user_id:
+        return json_response_not_owned(url_record.short_code)
+
     try:
         url = url_record.url
         db.delete(url_record)
         db.commit()
-        return JSONResponse(
-            {"message": f"deleted record {url_request.short_code} -> {url}"},
-            status_code=status.HTTP_200_OK)
+        return json_response_deleted(url_request.short_code, url)
 
     except Exception as e:
         db.rollback()
-        print
-        return RedirectResponse(
-            url=app.url_path_for("root"),
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        return json_response_failure()
 
 
 @app.delete("/delete_short_code/{short_code}")
 async def delete_url_short_code(
         short_code: str,
-        db: Session = Depends(get_db)):
+        db: Session = Depends(get_db),
+        session: SessionContainer = Depends(verify_session(session_required=False))):
+
+    user_id = get_user_id(session)
 
     url_record = db.get(ShortURLModel, (short_code))
 
     if url_record is None:
         return json_response_not_found(short_code)
+
+    if url_record.owner_id != user_id:
+        return json_response_not_owned(url_record.short_code)
+
     try:
         url = url_record.url
         db.delete(url_record)
         db.commit()
-        return JSONResponse(
-            {"message": f"deleted record {short_code} -> {url}"},
-            status_code=status.HTTP_200_OK)
+        return json_response_deleted(short_code, url)
 
     except Exception as e:
         db.rollback()
         print
-        return RedirectResponse(
-            url=app.url_path_for("root"),
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        return json_response_failure()
 
 
 @app.post("/modify_short_code")
 async def modify_url_short_code(
         mod_request: ModificiationRequest,
-        db: Session = Depends(get_db)):
+        db: Session = Depends(get_db),
+        session: SessionContainer = Depends(verify_session(session_required=False))):
+
+    user_id = get_user_id(session)
 
     if mod_request.short_code == '' or mod_request.new_short_code == '':
-        return JSONResponse(
-            {
-                "message": "you must supply a short code and a new short code"},
-            status_code=status.HTTP_406_NOT_ACCEPTABLE)
+        return json_response_missing("a short code and a new short code")
 
     url_record = db.get(ShortURLModel, (mod_request.short_code))
     conflict_url_record = db.get(ShortURLModel, (mod_request.new_short_code))
 
     if url_record is None:
         return json_response_not_found(mod_request.short_code)
+
+    if url_record.owner_id != user_id:
+        return json_response_not_owned(url_record.short_code)
 
     if conflict_url_record is not None:
         return json_response_in_use(mod_request.new_short_code)
